+++ packages/react-native-reanimated/apple/worklets/apple/WorkletsModule.mm
@@ -1,6 +1,6 @@
 #import <React/RCTBridge+Private.h>
 #import <worklets/WorkletRuntime/RNRuntimeWorkletDecorator.h>
-#import <worklets/apple/REAMessageThread.h>
+#import <worklets/apple/WorkletsMessageThread.h>
 #import <worklets/apple/WorkletsModule.h>
 #import <worklets/tools/SingleInstanceChecker.h>
 
@@ -36,7 +36,7 @@ RCT_EXPORT_BLOCKING_SYNCHRONOUS_METHOD(installTurboModule : (nonnull NSString *)
 {
   auto *bridge = self.bridge;
   auto &rnRuntime = *(jsi::Runtime *)bridge.runtime;
-  auto jsQueue = std::make_shared<REAMessageThread>([NSRunLoop currentRunLoop], ^(NSError *error) {
+  auto jsQueue = std::make_shared<WorkletsMessageThread>([NSRunLoop currentRunLoop], ^(NSError *error) {
     throw error;
   });
   nativeWorkletsModule_ = std::make_shared<NativeWorkletsModule>(std::string([valueUnpackerCode UTF8String]), jsQueue);
